phase-8-streamlit/requirements.txt
streamlit
datasets
python-dotenv
pandas
flask
pytest
